Add simple synchronization with skew limit
This adds a very simple synchronization mechanism to keep the secondary
video in sync with the primary video. On each primary video timeupdate
event the current time of the secondary video is compared to the time
from the timeupdate event. If the difference is too large the secondary
video will be seeked to time from the event.

As this invovled a few additional async code the synchronisation might
not be perfect. Edge cases and rate limiting are not handled if e.g.
getting the current time of the secondary video takes a longer amount of
time.