fix: chown build directory
The build container image we are using is running as the `user` user
(see Dockerfile) because running Chrome as root isn't support. The
chrome sandbox requires to be run as a user, or to be explicitly
disabled [1].

To still allow writing to the build directory as the `user` user, we
changed the owner of the build directory to `user:user`, but, with the
latest Gitlab Runner release, the build directory path changed.

Instead of cloning the repository into `/builds/gitlab/<group>/<name>`,
the new path is just `/builds/<group>/<name>`, as can be seen in the
pipeline output:

    Getting source from Git repository
    Fetching changes...
    Reinitialized existing Git repository in /builds/xikolo/video-player/.git/

The simplest solution is to adjust the `chown` call to use whatever the
current working directory, which always will be where the source was
checked out.

Another option would be to drop the custom user from the build base
image, and to call chrome with disabled sandboxing:

    // stencil.config.js
    testing: (() => {
      // CI runs in docker container as root, which isn't supported by any chrome
      // sandbox mode. We also do not need the sandbox because it runs a in
      // throwaway container and VM only running our tests.
      if (process.env.CI)
        return { browserArgs: ['--no-sandbox', '--disable-setuid-sandbox'] };

      // ...