TS: Improve signature for _invokePlayerFunction
By further splitting up the `XmVideo` interface (Interface Segregation
Principle, hello!), it is easy to type-hint the name of one of the
player functions.

After that, I only needed to explicitly list the generic types for the
call to `apply()` (in order: the type of `this` in the function, the
type of the optional `params`, and the function's return type).

I really like the outcome, as this lets us keep this handy utility
function, while also having more type safety because TypeScript will
prevent us from calling the helper with an invalid function name.

This is in preparation for enabling TypeScript's `noImplicitAny` option.