TS: Improve types for dealing with locales
We now have a type definition for a string that's one of our valid
locales, and also a friendly "type guard" function that can take a
string and check whether it is a valid `KnownLocale`, therefore avoiding
explicit type casts which are frowned upon in the TypeScript world.

This is in preparation for enabling TypeScript's `noImplicitAny` option.