Make screen component dumb
In order to switch between dual and single stream mode,
the Screen component needs to re-rendered.
So far, the reaction to the loading of the video ratio could only
be listened to once as soon as a Video component provides it.
This resulted in a collapsed layout of the screens.

Now, the player will provide these values for the Screen.