[XI-5568] Implement loading state and indicator in player
The loading state is initially true until the player is fully loaded.
Further state changes are based on two internal events
when video buffering starts and when buffering is finished.

These are mapped to the video provider events accordingly.

Vimeo has their own loading indicator implementation in the video.
We are forced to use it as we cannot override
the styles or logic in the embedded iframe.
So we remove buffering listeners from Vimeo provider and use theirs.