[Bugfix] Controls show a current time of 'NaN' while seeking
The payload of the 'seeked' event is undefined in the Kaltura API. This
caused the current time of the player to be NaN for a short time.

Moving the update of the seconds from the 'seeked' event listener to the
seek method, solves the issue.

For the same reason, seeking through the transcript was also broken for
Kaltura videos.

Closes XI-6068