Fix GitHub test action
With the puppeteer update to 19,
it stores browsers to globally cache browsers between installation.

This caused problems with our github actions,
because it could not find the cached version while running the tests.
leading to "[ ERROR ] Could not find Chrome" errors.

Instead, this commit reverts to version 18
which runs smoothly without limitations on our side.
The proposed configuration [1] to change the cache directory
did not work for our setup.

Skipping CI cache was not an sustainable option to me.

[1] https://pptr.dev/guides/configuration#changing-the-default-cache-directory